+++ packages/plugins/GeoLocation/src/utils/positionChanged.ts
@@ -1,0 +1,8 @@
+export default function positionChanged(
+  position: number[],
+  transformedCoords: number[]
+): boolean {
+  return (
+    position[0] !== transformedCoords[0] || position[1] !== transformedCoords[1]
+  )
+}
